Explicitly draft new sections with a nil change_note
When updating the first published at date the change note passed to the
section should be nil, this will then be skipped by the PublicationLogger
and avoid duplicate change notes appearing in the frontend.
Previously the section was copied along with the message in the change note
and these could appear several times in the updates page on the frontend.