Use Section#uuid in manual_helpers.rb
Note the introduction of the `entity_id_for` method. This is necessary because
it turns out that the
`check_manual_is_drafted_and_published_with_no_public_timestamps` and
`check_manual_was_published` methods are actually called with both `Manual`s
and `Section`s.

These method names should be changed to reflect the way that they're called but
that'll have to wait for another time.