Add unit spec for SearchIndexAdapter
As requested by @chrisroos.

I've chosen to only stub at the level of the Rummager API itself. Thus
this is actually more like an integration test at the moment, because
it's testing the behaviour of the combination of the following:

* `SearchIndexAdapter`
* `RummagerIndexer`
* `AbstractIndexableFormatter`
* `ManualIndexableFormatter`
* `SectionIndexableFormatter`

However, I'm hoping we can gradually collapse the last four classes into
the `SearchIndexAdapter`. So writing the tests like this is a bit of an
investment in making it easier to refactor in the future.

I've intentionally *not* stubbed the Rummager API using the
WebMock-based `GdsApi::TestHelpers::Rummager` helper methods, because
I think the latter is too low level and would unnecessarily duplicate
tests which should live in the `gds-api-adapters` gem.