Introduce SectionViewAdapter#id
This is required so that we can continue to call `form_for` and pass an
instance of our `SectionViewAdapter` to it once we remove `Section#id`. I've
had a quick look to see whether there's a more Railsy way of telling `form_for`
to use an alternative attribute to `id` but didn't find anything obvious.