Extract SearchIndexAdapter#indexable_manual & #indexable_section
I want to inline `RummagerIndex` into this class and I think doing this
will make that easier. Also I think this makes the code easier to
follow in its own right.

Note that the implementation of `indexable_section` wraps a section with
a `MarkdownAttachmentProcessor` when the section is going to be removed
from the index even though this is not necessary. I'm confident that
this has no effect on the overall behaviour because of the detailed
coverage provided by `SearchIndexAdapter` spec. I pretty confident the
extra level of indirection will not have any significant impact on
performance and, in any case, I expect the code in the formatters to be
collapsed into `SearchIndexAdatper` in subsequent commits.