Push down AbstractIndexableFormatter#indexable_attributes
I don't think that inheritance is the best way to remove the
duplication in the behaviour currently implemented in the formatter
classes. By pushing the implementation of this method down into the
sub-classes, we avoid the hard-to-follow Template Method pattern and
thus make the code easier to understand, even at the cost of some
duplication.

I'm working towards collapsing the formatter classes into
`SearchIndexAdapter` and hope to remove the duplication in other ways
in subsequent commits.