Add test coverage for AbstractIndexableFormatter#with_leading_slash
Ideally the format of the `Manual#slug` would be constrained by
validation and we wouldn't need the conditional logic in
`AbstractIndexableFormatter#link` which calls the method this commit is
testing. However, we're stuck with it for now. See #1041.

Also it would be simpler to test this in a test for the formatter
class itself. However, I'm in the process of collapsing the formatters
into `SearchIndexAdapter` and so testing it in here will make it easier
to refactor the code without changing the tests.