Rename #link -> #path in formatters
I think this better reflects the purpose of the method. Although I know
the name of the attribute send to the Publishing API is `link`, to me a
link sounds like an HTML anchor tag or at least a fully qualified URL
and not just a path which is what this is.