Introduce OrganisationsAdapter & Organisation model
See #996 for the motivations behind their introduction.

The main differences are that you now call `OrganisationsAdapter#find`
instead of `OrganisationFetcher#fetch`, and the former returns an
instance of the new `Organisation` model instead of `GdsApi::Response`.
The latter means we can avoid the duplicated logic which was obtaining
the various organisation attributes from difference bits of the response
hash.

I've added an `OrganisationsAdapter` spec which stubs at the
Organisations API level. This should make it easy to collapse
`OrganisationFetcher` into `OrganisationsAdapter` in a subsequent
commit. It's also consistent with what I've done for
`SearchIndexAdapter`.

I haven't added a unit spec for the `Organisation` model, because it
doesn't really have any behaviour. I had hoped to include
`ActiveModel::Model` to avoid writing `Organisation#initialize`, but
it isn't available in the current version of Rails.