Upgrade to Rails 4.2.8
I ran the following command in order to update the minimum set of
dependencies:

```
$ bundle update rails mongoid sprockets-rails sass-rails \
activemodel builder sprockets actionpack railties \
activesupport rack --conservative
```

Note that I had to remove the version lock of sass-rails so that it
could be upgraded.

Note that I've also had to upgrade Mongoid to 4.x as 3.x appears to be
tied to Rails 3.