Memo-ize adapters
c.f. the `Services` module.

This means that the same instance of each adapter is used for all
requests handled by a given process. In turn this means that the
`OrganisationsAdapter` cache is now actually useful!

There's no point in storing the underlying service instances in
instance variables in the adapters, because the former are themselves
memo-ized. Indeed storing the service instances in instance variables
led to problems in stubbing.