Replace SectionPresenter#fetch with named methods
This change makes it clearer what data the Presenter needs to return.

Note that there was already a `body` method on this presenter, I've
chosen to use the `attributes.fetch` form of the method definition for
consistency now and will refactor later.

This change revealed that the `updated_at` attribute in the
`SectionPublishingAPIExporter` spec wasn't used so it has been removed
and the test setup condensed.