Delegate SectionPresenter methods to `section`
After working through the pipeline of processers in
`SectionPresenter#call` I am fairly confident that the only property
of the `section` that is modified is the `body`. This means that all
of the other methods on this class can delegate directly to the
passed-in `section` without a change in behaviour.