Remove SectionHeaderExtractor and dependencies
The SectionHeaderExtractor added an attribute 'headers' to the passed
in Section in the SectionPresenter. This is the only place it was
called from. As SectionPresenter does not have a `headers` method in
its public interface as best as I can tell the SectionHeaderExtractor
and its dependency GovspeakHeaderExtractor are redundant. This commit
removes them.