Remove unnecessary tests from SectionPublishingAPIExporter spec
`SectionPublishingAPIExporter` is only ever constructed with
`update_type` set to either `:republish` or `nil`. The former is called
from the `Manual::RepublishService` via the
`PublishingApiDraftManualWithSectionsExporter`. The latter is the
default which is used in all other cases.

The `update_type` argument is a way to "override" the update type which
would be generated just from the state of the section. This generated
update type is only ever "major" or "minor".

The test I'm removing in this commit is unnecessary, because it
sets `update_type` (override) to "major" which never happens in
production.