Remove unnecessary tests from ManualPublishingAPIExporter spec
`ManualPublishingAPIExporter` is only ever constructed with
`update_type` set to either `:republish` or `nil`. The former is called
from the `Manual::RepublishService` via the
`PublishingApiDraftManualWithSectionsExporter`. The latter is the
default which is used in all other cases.

The `update_type` argument is a way to "override" the update type which
would be generated just from the state of the manual. This generated
update type is only ever "major" or "minor".

The tests I'm removing in this commit are unnecessary, because they set
the `update_type` (override) to "major" or "minor" which never happens
in production.