Extract Manual#all_sections_are_minor?
Although they're not brilliant, I've moved the relevant existing
examples from `ManualUpdateType` spec to `Manual` spec relatively
unchanged to keep the diff as simple as possible.

I've also had to change examples in `ManualPublishingAPIExporter` spec
to stub `Manual#all_sections_are_minor?`. This has then made a bunch of
the combinatorial examples redundant. I feel as if there is sufficient
test coverage in the examples for `Manual#all_sections_are_minor?` to
make it safe to do this.