Introduce specs for SectionPresenter
These specs tests the main functionality of SectionPresenter. I've
used an "integration" style of test to assert whether the rendered
HTML in the body has been processed in the expected way rather than
mocking the collaborators and asserting that they have been called.

The intention of introducing these specs at this stage is to make it
easier to refactor the SectionPresenter and inline some of the
functionality from its collaborators.