Process the body directly rather than the section
The various parts of the processing pipeline only affect the section
body. Before this commit the section object was passed into each
processor and the changes accumulated in an attributes hash. I think
it is conceptually simpler if we modify the body of the section
directly. This change uses the existing processors' `body` method to
do that which should make it easier to replace each one of these with
a more direct mutation of the body.