Inline `render_footnotes_heading` implementation
The purpose of the FootnotesSectionHeadingRenderer class is to perform
a simple substitution in the passed in body. This implementation can
be inlined directly in the method. As FootnotesSectionHeadingRenderer
is only called from this one place, it is also removed.

I'm satisfied that the spec added for ManualPresenter#body covers the
relatively simple behaviour of this code and there's no need to retain
the existing spec for FootnotesSectionHeadingRenderer.