Bring "lint" Rake task into line with "Lint Ruby" stage on CI
According to the task description this is supposed to mimic the "Lint
Ruby" stage run on Jenkins CI [1,2]. I've updated the `govuk-lint-ruby`
options and directories in the Rake task to more closely match those
used in the CI stage.

Note that this means we are now only checking code that has changed.
Ideally we'd be checking all the code every time, but if we want to do
that we should change the Jenkins CI build to do the same.

Note also that we are no longer checking the following directories:
`bin`, `config` & `features`. If we decide we want to do this then we
should chage the Jenkins CI build to do the same.

I did contemplate trying to change the Jenkins CI build, but there
doesn't seem to be an easy way to do this with the current
`govuk_jenkinslib.groovy` code.

[1]: https://github.com/alphagov/govuk-puppet/blob/43af3ededf1fc289d63bbac3969176bed76ce147/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L138-L140
[2]: https://github.com/alphagov/govuk-puppet/blob/43af3ededf1fc289d63bbac3969176bed76ce147/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L378-L396