Extract Manual#update_type
Strictly speaking the values returned from this method are specific to
the Publishing API. However, for the moment, given they're just
Strings, I can live without a "translation" between the domain model and
the Publishing API.

In `ManualPublishingAPIExporter` spec, I'm now stubbing the new method
which renders more of the combinatorial explosion of examples redundant.

Although they're not perfect, I've moved the examples from the
`ManualUpdateType` spec into the `Manual` spec changing them as little
as possible in the process.