Extract Section#update_type
Strictly speaking the values returned from this method are specific to
the Publishing API. However, for the moment, given they're just
Strings, I can live without a "translation" between the domain model and
the Publishing API.

In `SectionPublishingAPIExporter` spec, I'm now stubbing the new method
which renders more of the combinatorial explosion of examples redundant.

I've added coverage for the new method in the `Section` spec.