Convert Manual#update_type -> version_type
The new method returns one of the following symbols:

* `:new`
* `:minor`
* `:major`

These are then mapped to the relevant Publishing API `update_type`
strings in `ManualPublishingAPIExporter#update_type`.

This means that `Manual#version_type` does not contain logic specific
to the Publishing API. Also I think it better captures the fact that
"new" versions of a document should always be sent to the Publishing API
with the `update_type` set to "major".

This change has allowed me to simplify `ApplicationHelper#publish_text`
a little, because we are now distinguishing between a "new" version of
a manual and a "major" version.

I've decided to remove the comments that were in `Manual#update_type`,
because I think their content is now better captured in the code and in
the specs.