Convert Section#update_type -> version_type
The new method returns one of the following symbols:

* `:new`
* `:minor`
* `:major`

These are then mapped to the relevant Publishing API `update_type`
strings in `SectionPublishingAPIExporter#update_type`.

This means that `Section#version_type` does not contain logic specific
to the Publishing API. Also I think it better captures the fact that
"new" versions of a document should always be sent to the Publishing API
with the `update_type` set to "major".

It also means `Section#version_type` is now consistent with
`Manual#version_type`.

I've decided to remove the comment that were in `Section#update_type`,
because I think its content is now better captured in the code and in
the specs.