Extract PublishingAdapter#save
I've written the spec for `PublishingAdapter` from scratch by inspecting
the code. As with the `SearchindexAdapter` spec, it's more of an
integration test in that it tests the combination of the various
exporter classes and only stubs at the Publishing API level. The idea is
that this will make it easier to collapse the exporter code into
`PublishingAdapter#save` without having to change the tests.

Contrary to the `SectionPublishingAPIExporter` spec, I've chosen to use
real instances of the various model classes, rather than mock objects.
I think this makes the examples more representative and therefore more
likely that the spec will catch problems caused by changes to these
model classes.

Contrary to the `SectionPublishingAPIExporter` spec, I've chosen not to
provide a `content_id` attribute for the `Attachment` instance, because
the class does not have a field with this name. As such, I believe that
in production the `content_id` for an `Attachment` is always obtained
anew from `SecureRandom.uuid`. Thus I think the example in the new
`PublishingAdapter` spec is more representative of what happens in
production. See #1082 for more details.

Note that I've made an `excluding` RSpec matcher available by inverting
the built-in `include` matcher.