Replace links schema check from exporter spec in adapter spec
I want to inline `ManualPublishingAPILinksExporter` into
`PublishingAdapter` and doing this should make that easier, because the
new spec is less implementation-specific.

It might've been neater to change the `Section#uuid` and
`Organisation#content_id` across the whole of the `PublishingAdapter`
spec, but I think it's useful to have the more expressive strings,
"section-uuid" and "organisation-content-id" for most of the examples.
I'd hope that eventually we'll do some more refactoring and make a
better test for this behaviour, so I'm not overly bothered that this
spec looks a bit clunky.