Use boolean keyword arg for PublishingAdapter#save
Convert `action` string argument into `republish` boolean keyword
argument. The `action` argument was always either "republish" or `nil`.
Making it a boolean makes this more explicit and making it a keyword
argument makes the code more readable.

I'm also intending to add another "option" to this method and I think
that'll be easier with the new method signature.