Introduce include_sections option to PublishingAdapter#save
And use it to replace calls to `PublishingApiDraftManualExporter#call`.
This exporter was doing the same thing as `PublishingAdapter#save`, but
only sending links & content to the Publishing API for only the manual,
not its sections as well. By introducing this option, we can get rid of
the `PublishingApiDraftManualExporter` class altogether.