DRY up PublishingAdapter using #save_section from #save
I've added an `update_type` option to `#save_section` to support this.
Note that the default value for `update_type` in the
`SectionPublishingAPIExporter` constructor is `nil`, so this doesn't
change the behaviour of `#save_section` when `update_type` is not
supplied.

Also there shouldn't be any deterioration in performance due to looking
up the organisation for each section being exported. This is because
`OrganisationsAdapter` caches the organisation anyway.