Extract PublishingAdapter#save_manual_links
To improve readability. As before there shouldn't be a performance hit
even though we're now looking up the organisation every time we export
the links for a manual as well as elsewhere, because the organisation
will be cached in `OrganisationsAdapter`.

Note that I'm planning to extract other methods along similar lines.