Extract PublishingAdapter#save_manual_content
To improve readability. As before there shouldn't be a performance hit
even though we're now looking up the organisation every time we export
the content for a manual as well as elsewhere, because the organisation
will be cached in `OrganisationsAdapter`.

I'm expecting this method to get bigger as I inline logic from
`ManualPublishingAPIExporter`.

Note that I'm planning to extract other methods along similar lines.