Extract PublishingAdapter#save_section_content
To improve readability. As before there shouldn't be a performance hit
even though we're now looking up the organisation every time we export
the content for a section as well as elsewhere, because the organisation
will be cached in `OrganisationsAdapter`.

I'm expecting this method to get bigger as I inline logic from
`SectionPublishingAPIExporter`.