Introduce include_links option to PublishingAdapter#save
And use it to replace calls to `ManualPublishingAPIExporter#call` and
`SectionPublishingAPIExporter#call` from `ManualRelocator#send_draft`.

Although it's not obvious to me whether it was intentional, the only
difference in what `send_draft` was doing compared to
`PublishingAdapter#save` was doing was that the former was not exporting
the manual or section links to the Publishing API. The new option allows
exporting links to be disabled.

This removes the only reference to `ManualPublishingAPIExporter` and
`SectionPublishingAPIExporter` outside `PublishingAdapter`. Thus this
change should make it easier to inline these exporters into the adapter.

Note that one other difference is that `ManualRelocator#send_draft` did
not check whether each section needed exporting. Again it's not obvious
whether this is intentional. However, it looks as if the test coverage
for `ManualRelocator` is pretty comprehensive and all the examples pass,
so I'm happy that it's safe to make this change. I've added #1083 to
suggest we investigate whether `ManualRelocator#send_draft` should
be exporting links for the manual and its sections after all.