Move section links schema check into PublishingAdpater spec
I want to inline `SectionPublishingAPILinksExporter` into
`PublishingAdapter` and this should make that easier, because it's less
implementation-specific.

It might've been neater to change the `Manual#id` and
`Organisation#content_id` across the whole of the `PublishingAdapter`
spec, but I think it's useful to have the more expressive strings,
"manual-id" and "organisation-content-id" for most of the examples.
I'd hope that eventually we'll do some more refactoring and make a
better test for this behaviour, so I'm not overly bothered that this
spec looks a bit clunky.