Pass Section vs SectionEdition to PublishingAPIRedirecter#call
It feels more natural/consistent to pass in an instance of `Section`. I
suspect the only reason we were previously passing in an instance of
`SectionEdition` was that we didn't have an instance of `Section` and it
wasn't straightforward to build one. However, since this commit [1] we
have had an instance of `Section` available.

Note that the only method called on the `entity` passed into
`PublishingAPIRedirecter#call` is `#slug` which `Section` delegates to
the latest `SectionEdition`, so this change should not cause any change
in behaviour.

This change will make it easier for me to extract
`PublishingAPIRedirecter#call` into a suitable high-level method on
`PublishingAdapter`.

[1]: 188db8e5f8ce55f7a709b799e20d24bc948c271d