Extract PublishingAdapter#redirect_section
I've added the usual integration-y test to the `PublishingAdapter` spec
which will make it easier to refactor in subsequent commits.

Technically, the `PublishingAPIRedirecter` could be used to redirect
any piece of content managed by the Publishing API, i.e. either a manual
_or_ a section. However, as far as I can see it's only ever called from
the `SectionReslugger` where it is used to redirect a section. This is
why I decided to name the method `#redirect_section` and not just
`#redirect`. I decided to abbreviate the `redirect_to_location` argument
name to `to`, in `PublishingAdapter#redirect_section`, because I think
it makes the calls terser and more readable.