Extract CHILD_SECTION_GROUP_TITLE constant into GdsApiConstants::PublishingApiV2
This makes the code more explicit and removes the duplication between
the app and the specs.

Since this value is sent to the Publishing API in the `details` Hash, I
think it effectively passes through to the Content Store unchanged. Thus
it's value is probably only of significance to the Manuals Frontend
app. It would be nice if this constant could be shared between the apps
somehow.