Extract EDITION_LOCALE constant into GdsApiConstants::PublishingApiV2
This makes the code more explicit and removes some duplication in the
app as well as between the app and the specs.

According to the Publishing API docs [1] this value represents the
locale of the edition being sent to the API. At the moment this is
hard-coded to "en", but I could imagine it being obtained more
dynamically from the user, e.g. if they wanted to publish something in
Welsh...?

[1]: https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-v2contentcontent_id