Extract UPDATES_PATH_SUFFIX constant into GdsApiConstants::PublishingApiV2
This makes the intention of the code a bit more explicit and removes
some duplication between the application and the specs.

As far as I can see, although this route path is sent to the Publishing
API for a manual [1], it doesn't appear in either the publishing [2] or
frontend [3] examples for a manual in the content schema repo. What's
more it doesn't appear to be used by the Manuals Frontend app, even
though that app renders a link [4] to an "updates" page with this path.

I could imagine it might have been intended to be used for search
indexing, but the Manuals Publisher app is still indexing directly into
Rummager, so I don't think that can be true at the moment.

I've opened #1088 to remind us to investigate whether we need to send
this route at all when we are sending a manual to the Publishing API.
If we don't we could remove it and simplify the code.

[1]: https://github.com/alphagov/manuals-publisher/blob/e413d4a5ab31a9591f436eacefb1732d7bc5ed60/app/exporters/manual_publishing_api_exporter.rb#L53
[2]: https://github.com/alphagov/govuk-content-schemas/blob/e01021233c8ec772eaad3144fb6bb111925a6788/formats/manual/publisher_v2/examples/manual.json#L79-L83
[3]: https://github.com/alphagov/govuk-content-schemas/blob/e01021233c8ec772eaad3144fb6bb111925a6788/formats/manual/frontend/examples/content-design.json
[4]: https://github.com/alphagov/manuals-frontend/blob/371b3932dc0ff34cad20659661a6677820761ee0/app/presenters/manual_presenter.rb#L18-L20