Don't swallow GdsApi exceptions in PublishingAdapter#discard_section
As far as I can see this is method only ever called from
`Section::RemoveService#discard_section_via_publishing_api`. This method
is only called as part of the `SectionsController#destroy` action which
is only accessible from the page rendered by the
`SectionsController#withdraw` action. The latter action makes use of the
`Section::ShowService` which will only work if a section has been
created and consequently sent to the Publishing API. I think this means
that we can expect that in normal operation these `GdsApi` exceptions
will not be raised and if they are, we want to know about them. Thus I'm
happy to remove this `rescue` block.