Stop checking update_type in PublishingAPIPublisher
`PublishingAPIPublisher#call` is only ever called from
`PublishingApiManualWithSectionsPublisher#call`. The latter only ever
calls the former with an `update_type` of "republish" or `nil`, both of
which are allowed by `PublishingAPIUpdateTypes#check_update_type!`. Thus
it doesn't feel as if this check is adding much value and it feels
distinctly non-idiomatic. What's more removing it will make it easier
for me to inline `PublishingAPIPublisher#call` into
`PublishingApiManualWithSectionsPublisher#call`.

If we decide that this check is important, I think the right place for
it is probably in `GdsApi::PublishingAPIv2` and this kind of check
should then be applied consistently for all arguments of all methods on
the Publishing API.