Don't swallow GdsApi exceptions in PublishingAdapter#publish
I can't see how this exception would happen in normal operation other
than perhaps if there is a double form submission. I think we're better
off removing this `rescue` block and fixing the underlying problem
if/when it arises. Note that we will be notified via Errbit if such an
exception occurs so I'm happy to make this change.