Pass attachment vs attributes to attachment_json_builder method
In general I think it's better to pass around models, rather than IDs or
hashes of attributes. While this temporarily makes the code more
verbose, I'm intending to change the calls to `Hash#fetch` on the
attributes to calls to attribute reader methods which will make the code
terser again.