Remove SectionPublishingAPIExporter spec
The behaviour covered in this spec is all now covered by the
`PublishingAdapter` spec albeit in a more integration-y form.

In hindsight, I should probably have removed it when I added the
examples to the `PublishingAdapter` spec.

I'm removing it now in order to facilitate refactoring
`SectionPublishingAPIExporter` with a view to in-lining it into
`PublishingAdapter`.