Upgrade gds-api-adapters to v45.0.0
I've read through the `CHANGELOG` between v39.2.0 and v45.0.0 and I
couldn't see anything that looked worrying. What's more all the tests
pass after this upgrade.

Note that I've intentionally avoided upgrading `rest-client` to its
latest version (v2.0.2), because it introduced warnings when I ran the
tests and the new version of `gds-api-adapters` does not require it to
be upgraded from v2.0.0.