Remove unnecessary calls to Time#iso8601
There's no need to convert instances of `Time` into an ISO8601 string
when we're sending a Hash to `GdsApi::PublishingApiV2`, because
`gds-api-adapters` does this in `GdsApi::JsonClient#do_json_request` [1]
when `Object#to_json` gets called on the `params` Hash passed in.

Note that `JSON.parse` does not reverse this conversion, so I've left
the relevant test assertions in `ManualHelpers` unchanged. I've also
left `DateTimeHelper#nice_time_format`, because that is unrelated to
exporting to the Publishing API.

[1]: https://github.com/alphagov/gds-api-adapters/blob/master/lib/gds_api/json_client.rb#L137