Use set(hash) instead of set(key, value)
This new syntax was introduced in Mongoid 4[1]:

> Document#set now accepts multiple attributes in the form of a hash,
> instead of the previous (field, value) args. Field aliases and
> typecasting are also now supported in this operation.

Note that the manual_relocator_spec was already using `#set(hash)` in
places but that wouldn't have been doing what we were expecting before
the upgrade to Mongoid 4. It would've been creating a field whose name
is a string representation of the hash.

[1]: https://github.com/mongodb/mongoid/blob/master/CHANGELOG.md