Always add attachments key to details hash in section exporter
Even if it's an empty array.

As far as I can see this `attachments` value is not used in the
Manuals Frontend app, even though it does appear in the content schema
frontend example for a section [1].

This makes sense, because we send a rendered HTML version of the section
body to the Publishing API which already incorporates links to the
attachments. This HTML version takes priority over the Govspeak version
which is also sent.

Even if the latter was being used, the Publishing API app passes such
an `attachments` value as an option to `Govspeak::Document.new` [2].
This value is converted to an array using `Array.wrap` in the `govspeak`
gem [3].

So all in all, I can't see a problem with passing an empty array of
attachments to the Publishing API. In fact, if anything, it seems better
and it makes the code in `SectionPublishingAPIExporter#call` simpler.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/30e883a250416b56af5d0e96ac0b100752f25a20/formats/manual_section/frontend/examples/what-is-content-design.json
[2]: https://github.com/alphagov/publishing-api/blob/47ecc7f10829580e32f2b4c1e7de5e98af34f92a/app/presenters/details_presenter.rb#L63
[3]: https://github.com/alphagov/govspeak/blob/34c5a303241debc52248ca0eb20eddca3f89cac2/lib/govspeak.rb#L38