Pass version_type vs update_type into section exporter constructor
Passing in a version type of either `:republish` or `nil` allowed me to
simplify `SectionPublishingAPIExporter#update_type`, because I was able
to replace the guard condition with an extra `when` clause in the `case`
statement.

The existence of both `PublishingAdapter#update_type` *and*
`#version_type` is sub-optimal, but I'm confident I will be able to do
something similar with `ManualPublishingAPIExporter` which will mean
I'll be able to get rid of `PublishingAdapter#update_type` at that
point.