Use as_json instead of iso8601 in manual_helpers.rb
This avoids test failures in the "Creating a manual that was previously
published elsewhere" scenario of creating-and-editing-a-manual.feature.

The problem is due to a change in the way Rails coerces time to a string
for JSON encoding. It now includes milliseconds by default where
`#iso8601` doesn't[1].

The failing tests appear to be checking the data that's being sent from
the gds-api-adapters Gem to the Publishing API. We know that the
gds-api-adapters Gem is using `to_json` to serialise this data so we can
use that to ensure that our timestamps match. See PR #1097 for some
related information.

[1]: https://github.com/rails/rails/pull/9128