Update config/environments/production.rb
I ran `rake rails:update` to generate the updated config files.

The `config.serve_static_files` option was previously hardcoded to
false. We're happy to be accept the new Rails 4.2 default of controlling
it via RAILS_SERVE_STATIC_FILES as that seems to be the case in lots of
other apps[1].

The `config.assets.version = "1.0"` option is now configured in
config/initializers/assets.rb.

The `config.secret_token = ENV["SECRET_TOKEN"]` is now configured in
secrets.yml.

The `config.log_level` option has been changed from `:info` to `:debug`.
It looks as though it was only set to `:info` as that was the default
when the app was created. As such we're happy to accept this new default
of `:debug`.

Precompiling additional assets is now handled in assets.rb initializer.

[1]:
https://github.com/search?p=5&q=org%3Aalphagov+RAILS_SERVE_STATIC_FILES&type=Code