Fix section publishing when relocating a manual
When moving a manual from one URL to another, ensure that its sections
are properly created as drafts with the correct slug and then published.

Previously, the section publishing step was failing because the section
drafts had not been created. This was because the `PublishingAdapter`
only creates drafts for sections if the `republish` flag is added or if
the section has never been exported. This has been fixed by adding the
`republish` flag.

The tests did not catch the problem because the sections in the test
data had no export date, so they *were* being created as drafts and
republished. This has been fixed by adding export dates to the sections
in the test.