Stop checking update_type in ManualPublishingAPIExporter
`ManualPublishingAPIExporter#call` is only ever called from
`PublishingAdapter#save_manual_content` and in turn only from
`PublishingAdapter#save_manual`. Thus
`ManualPublishingAPIExporter#call` is only ever called with an
`update_type` of "republish" or `nil`, both of which are allowed by
`PublishingAPIUpdateTypes#check_update_type!`. Thus it doesn't feel as
if this check is adding much value and it feels distinctly
non-idiomatic. What's more removing it will make it easier for me to
inline `ManualPublishingAPIExporter#call` into
`PublishingAdapter#save_manual_content`.

If we decide that this check is important, I think the right place for
it is probably in `GdsApi::PublishingAPIv2` and this kind of check
should then be applied consistently for all arguments of all methods on
the Publishing API.

c.f. fb65f3852a4d6495fedb0e02e48d30003b95a502