Pass version_type vs update_type into manual exporter constructor
Passing in a version type of either `:republish` or `nil` allowed me to
simplify `ManualPublishingAPIExporter#update_type`, because I was able
to replace the guard condition with an extra `when` clause in the `case`
statement.

The existence of both `PublishingAdapter#update_type` *and*
`#version_type` is sub-optimal, but I'm hopeful there's some more
simplification that can be done around that.