Make args for delete_draft_manual Rake task optional
In this commit [1] both these arguments were made into required
arguments. However, the underlying `CliManualDeleter` class only expects
one or the other to be set. This change made it awkward to use the
Rake task to delete manuals using only a slug, because the Rake task
was forcing you to supply an ID which it uses in preference to the slug.

By making both Rake task arguments optional, we can delegate
responsibility for argument checking to the underlying class and thus
make it easier to delete a manual using only a slug.

[1]: 243ea72233759382939d13b94964d256c330db09