Use a Manual vs ManualRecord in CliManualDeleter#complete_removal
This means we can use a higher-level of abstraction. Ideally
`ManualRecord` should be an implementation detail of `Manual`.

I've also introduced `Manual#destroy` to keep the code as simple as
possible and at the same level of abstraction.

I've removed the now unused `CliManualDeleter#section_uuids` method.

And I've removed the large block comment, because I'm not sure it was
adding much and I think it might've been slightly out-of-date.