Exclude arbitrary attrs when copying SectionEditions
This should hopefully avoid the following errors when creating new
drafts of SectionEditions that contain old data:

> Mongoid::Errors::UnknownAttribute: Problem: Attempted to set a value
> for '<old-field-name>' which is not allowed on the model SectionEdition.
> Summary: Without including Mongoid::Attributes::Dynamic in your model
> and the attribute does not already exist in the attributes hash,
> attempting to call SectionEdition#old-field-name= for it is not
> allowed. This is also triggered by passing the attribute to any method
> that accepts an attributes hash, and is raised instead of getting a
> NoMethodError. Resolution: You can include
> Mongoid::Attributes::Dynamic if you expect to be writing values for
> undefined fields often.

We've been seeing this error since the upgrade to Mongoid 4 where the default
behaviour of allowing dynamic fields has changed. In Mongoid 3 the
behaviour was controlled by `Mongoid.allow_dynamic_fields` which was set
to true by default. In Mongoid 4 it's controlled at the model level by
including the `Mongoid::Attributes::Dynamic` module.

I think the test I've added could be better but I think it's good enough
for now. I have a separate branch where I'm refactoring Section and I
hope to improve the tests in there.

Note the call to `#with_indifferent_access` in the setup for the test.
This better mirrors the behaviour of `SectionEdition#attributes`.

The fields being copied are the fields listed in the `SectionEdition`
model, plus `_id`, plus `created_at` and `updated_at` from `include
Mongoid::Timestamps`,  minus the embedded `attachments`. The embedded
attachments are manually copied in the `Section#new_draft` method
although a further improvement might be to copy them directly from the
`attributes` hash.  Note that the `_id`, `created_at`, `updated_at` and
`exported_at` attributes are ignored as they appear in the
`no_copy_attributes` array. I plan to improve that in a subsequent
commit.