Inline Section#never_published?
Note that the name of this method was slightly confusing. I think it
implies that it inspects all editions to see whether the section has
ever been published whereas it actually only looks at the editions passed to
`Section#initialize`. We only ever retrieve Sections by using
`Section.find` which only supplies the two most recent editions to the
constructor.