Extract Manual.find_by_slug!
I'm planning to use this to DRY-up the code in a number of places where
a `ManualRecord` is looked up in order to build the relevant `Manual`.

It should also allow me to change a bunch of code to operate on
instances of `Manual` vs `ManualRecord` which is desirable, because the
latter should just be an implementation details of the former.