Use set(hash) instead of set(key, value) in ManualRelocator
This was missed in this commit [1] and has not been caught since the
upgrade to Mongoid v4, because the spec did not exercise this line of
code.

I exposed this problem when I split up the examples in `ManualRelocator`
spec. I decided to split up the examples, because I was running into
problems with refactoring `ManualRelocator`, because these examples were
somehow relying on the state of the `ManualRelocator` instance,
i.e. `subject`.

I don't think reusing the `ManualRelocator` instance is a very realistic
use case, because the `relocate_manual` Rake task calls
`ManualRelocator.move` which instantiates a new `ManualRelocator` each
time.

By splitting these examples up, a new `ManualRelocator` is instantiated
in each example.

[1]: a514f6e6a3c58db187405e2eec9fa7899b4658ff